+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 New features and improvements:
 * ...
 
+Other changes:
+* Changed dependencies to dataclasses (instead of attrs) and tomli (instead of toml) (#362)
+
 
 #### 1.8 (2021-11-25)